Extract abstract class TextField
